Added lost connection detection in run_evil()
If evil doesn't receive nothing from good in a second, it stops.
Removed game size input from command line, it can be useful only if it's
possible to send this information to the opponent, otherwise it's useful only
in offline mode.
Fixed order in sketcher.cpp function implementation, now it's consistent with
sketcher.h declaration.